pdf_querying_summarization/requirements.txt
accelerate==0.20.3
aioboto3==11.2.0
aiobotocore
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
asttokens==2.2.1
async-timeout==4.0.2
atlassian-python-api==3.39.0
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bitsandbytes==0.39.0
blinker==1.6.2
boto3
botocore
Brotli==1.0.9
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
chromadb==0.3.26
click==8.1.3
clickhouse-connect==0.6.4
cmake==3.26.4
coloredlogs==15.0.1
comm==0.1.3
cryptography==41.0.1
dataclasses-json==0.5.9
datasets==2.12.0
debugpy==1.6.7
decorator==5.1.1
deeplake==3.6.10
Deprecated==1.2.14
dill==0.3.6
duckdb==0.8.1
einops==0.6.1
entrypoints==0.4
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.99.1
filelock==3.12.1
flatbuffers==23.5.26
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
graphlib-backport==1.0.3
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
html2text==2020.1.16
httptools==0.5.0
huggingface-hub==0.15.1
humanfriendly==10.0
humbug==0.3.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==6.0.0
inflate64==0.3.1
InstructorEmbedding==1.0.1
ipykernel==6.23.1
ipython==8.12.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter-client==8.2.0
jupyter-core==5.3.0
langchain==0.0.221
langchainplus-sdk==0.0.19
lit==16.0.5.post0
llama-hub==0.0.9
llama-index==0.7.4
loralib==0.1.1
lz4==4.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
multivolumefile==0.2.3
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numcodecs==0.11.0
numexpr==2.8.4
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==2.0.2
parso==0.8.3
pathos==0.3.0
pdfminer.six==20221105
pdfplumber==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
platformdirs==3.5.3
posthog==3.0.1
pox==0.3.2
ppft==1.7.6.6
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.2.0
pure-eval==0.2.2
py7zr==0.20.5
pyarrow==12.0.0
pybcj==1.0.1
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.10.10
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.7.0
Pympler==1.0.1
pypdf==3.12.0
pyppmd==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.1.0
pyzstd==0.15.9
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
retrying==1.3.4
rich==13.4.2
rpds-py==0.8.10
s3transfer
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
stack-data==0.6.2
starlette==0.27.0
streamlit==1.24.1
sympy==1.12
tenacity==8.2.2
texttable==1.6.7
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
triton==2.0.0
trl==0.4.4
typing-extensions==4.5.0
typing-inspect==0.8.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
Wand==0.6.11
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
wikipedia==1.4.0
wrapt==1.15.0
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
